Update pymupdf_rag.py to fix compatibility issue with Python 3.9
The code in the pdf4llm library used the | operator for type hinting, which is only supported in Python 3.10 and above. This caused a TypeError when running the code on Python 3.9.

To resolve this issue, the code has been modified to use a compatible syntax that works with Python 3.9. The changes ensure that the library can be used with Python 3.9 without encountering the TypeError.